YUK-17111 IVVC Simulator SetPoint support (2) Fixed Set Point math. Changed Set Point to automatically update voltage always, not just when a new Set Point value is received from CapControl. Added Set Point unit test.

+++ yukon-client/simulators/src/com/cannontech/simulators/dao/RegulatorEventsSimulatorDao.java
@@ -12,7 +12,7 @@ public interface RegulatorEventsSimulatorDao {
         public Integer regulatorId;
         public RegulatorEvent.EventType eventType;
         public Instant timeStamp;
-        public Float setPointValue;
+        public Double setPointValue;
     }
     
     List<RegulatorOperations> getRegulatorEventOperationsAfter(Instant lastRegulatorEvaluationTime);
